Add rich Slack previews for Stash links
