+++ frontend/src/app/(app)/stashes/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 
-import { SSR_BACKEND_ORIGIN as BACKEND_ORIGIN } from "@/lib/backendOrigin";
+import { metadataForPublicStash } from "@/lib/stashMetadata";
 
 import StashPageClient from "./StashPageClient";
 
@@ -10,39 +10,20 @@ export async function generateMetadata({
   params: Promise<{ slug: string }>;
 }): Promise<Metadata> {
   const { slug } = await params;
-  const data = await loadPublicStash(slug);
-  if (!data) return { title: "Stash · Stash" };
-  const title = `${data.stash.title} · Stash`;
-  const description =
-    data.stash.description ||
-    `A Stash of ${data.items.length} item${data.items.length === 1 ? "" : "s"} from ${data.workspace_name}.`;
+  const metadata = await metadataForPublicStash({
+    slug,
+    path: `/stashes/${slug}`,
+  });
   return {
-    title,
-    description,
+    ...metadata,
     alternates: {
-      canonical: `/stashes/${slug}`,
+      ...metadata.alternates,
       types: {
         "text/markdown": `/stashes/${slug}.md`,
         "application/json": `/stashes/${slug}.json`,
       },
     },
-    openGraph: {
-      title,
-      description,
-      type: "article",
-      url: `/stashes/${slug}`,
-      siteName: "Stash",
-    },
-    twitter: { card: "summary_large_image", title, description },
   };
-}
-
-async function loadPublicStash(slug: string) {
-  const res = await fetch(`${BACKEND_ORIGIN}/api/v1/stashes/${slug}`, {
-    cache: "no-store",
-  });
-  if (!res.ok) return null;
-  return res.json();
 }
 
 export default async function StashPage({
